backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.24.1
beautifulsoup4==4.12.2
bs4==0.0.1
email-validator==2.0.0
sqlalchemy==2.0.20
aiohttp==3.8.5
asyncio==3.4.3
python-multipart==0.0.6
jinja2==3.1.2
APScheduler==3.10.4
python-telegram-bot==20.6
PyJWT==2.8.0
selenium==4.15.2
PyPDF2==3.0.1
python-docx==1.0.1
google-ads==22.1.0 